moved all the files into the src folder
