+++ src/data_models/MetaFeatures.py
@@ -1,5 +1,5 @@
 from app import db, ma
-from constants import TABLE_META_FEATURES
+from utils.constants import TABLE_META_FEATURES
 
 
 # Meta Feature Class/Model
