+++ src/data_models/Metrics.py
@@ -1,7 +1,7 @@
 from app import db, ma
 from data_models.Params import ParamSchema
 from data_models.MetaFeatures import MetaFeatureSchema
-from constants import TABLE_METRIC, CLASS_PARAM, CLASS_META_FEATURE
+from utils.constants import TABLE_METRIC, CLASS_PARAM, CLASS_META_FEATURE
 
 # Metric Class/Model
 class Metric(db.Model):
